reorganizing modal functionality into CommentsSection/index.js
