+++ frontend/src/components/StoryPage/LoggedInStory.js
@@ -1,22 +1,13 @@
-import React, { useState } from "react";
+import React from "react";
 
 import CommentsSection from "../CommentsSection";
-import { CommentModal } from "../../context/Modal";
 
 const LoggedInStory = ({ content, storyId }) => {
-	const [openModal, setOpenModal] = useState(false);
-
-	const onClose = () => setOpenModal(!openModal);
 
 	return (
 		<>
 			<p className="logged-in-story-page-content">{content}</p>
-			<button onClick={onClose}>Reply</button>
-			{openModal && (
-				<CommentModal onClose={onClose}>
-					<CommentsSection storyId={storyId} />
-				</CommentModal>
-			)}
+			<CommentsSection storyId={storyId} />
 		</>
 	);
 };
